Add appointment calendar to user pages
* calendar view toggle, appointment css
* Add faker to Gemfile
* Add appointment factory
* Add month_link helper method
* share ActiveRecord connection between RSpec and Capybara

We're using transactional fixtures to help keep our data clean. This
means that RSpec and Capybara are working in separate connections, and thus
separate transactions. This means that data created in Rspec isn't available
to capybara until the transaction is committed, which doesn't ever
happen. To get around this, we can share the connection between RSpec
and Capybara. Jose Valim created the solution we used here.

Here is a link to the SO article
http://stackoverflow.com/questions/8774227/why-not-use-shared-activerecord-connections-for-rspec-selenium

and also the original Jose Valim article
http://blog.plataformatec.com.br/2011/12/three-tips-to-improve-the-performance-of-your-test-suite/